Defines instructions and criterias for the task
This is our tech task for Flutter. You will be required to fork this
project and read through this README to understand what the use story is
and fulfil all the criterias listed there.

We hope you enjoy doing this task and do it with all the developmen best
practices you collected throughout your carrier.

Here a some tips for those who read commit messages:
- Do it using OOP, and using classes and methods alone is not OOP.
- You need to include tests. Even if you don't have the habit to do TDD
you are required to know how to create tests that cover all the user
story criteria.
- Do atomics and good commit messages, if you don't know it, a quick
research about it will help you get extra points.
- A good understanding of SOLID and how you organise and isolate the
business logic are important.

We can't wait to have you on-board. Good luck!